Restructured to allow compiling to both binary and WASM targets. WASM-bindgen not yet set up.
