+++ rust/Cargo.toml
@@ -1,9 +1,21 @@
 [package]
 name = "folkfriend"
 version = "3.0.0"
+authors = ["Tom Wyllie <tom@wyllie.dev>"]
 edition = "2018"
+description = "Transcription and recognition of traditional instrumental folk music"
+documentation = "https://github.com/TomWyllie/folkfriend/blob/master/README.md"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+name = "folkfriend"
+path = "src/lib.rs"
+crate-type = ["cdylib", "rlib"]
+
+[[bin]]
+name = "folkfriend"
+path = "src/bin.rs"
 
 [dependencies]
 clap = "3.0.0-beta.4"
@@ -17,4 +29,5 @@ reqwest = { version = "0.11.4", features = ["blocking"] }
 rustfft = "6.0.1"
 serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0"
+wasm-bindgen = "0.2"
 wav = "1.0.0"
