+++ rust/binary_build.sh
@@ -1,0 +1,1 @@
+cargo build --release
