+++ rust/wasm_build.sh
@@ -1,0 +1,1 @@
+wasm-pack build --target web
